Refactor code to add create helper functions for card elements
