+++ js/helpers/create/createCardBodyContainer.mjs
@@ -1,0 +1,5 @@
+export function createCardBodyContainer() {
+    const cardBody = document.createElement("div");
+    cardBody.classList.add("card-body", "px-1", "m-0");
+    return cardBody;
+}
