+++ js/helpers/create/createIconElement.mjs
@@ -1,0 +1,6 @@
+// Helper function to create an icon element with the given class
+export function createIconElement(iconClass) {
+    const icon = document.createElement('i');
+    icon.className = iconClass;
+    return icon;
+}
