+++ js/helpers/create/createImgWrapperContainer.mjs
@@ -1,0 +1,5 @@
+export function createImgWrapperContainer() {
+    const imgWrapper = document.createElement("div");
+    imgWrapper.classList.add("position-absolute", "w-100", "h-100", "top-0", "start-0");
+    return imgWrapper;
+}
